backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Validation & settings
pydantic==2.7.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# HTTP client (for Ollama)
httpx==0.27.0

# PDF parsing
PyMuPDF==1.24.1

# Vector embeddings
sentence-transformers==2.7.0

# Vector database
qdrant-client==1.9.1

# PDF report generation
reportlab==4.1.0

# Environment variables
python-dotenv==1.0.1
